Update Sample to Set Temperature and Voltage settings for Lithium Charge Profile
Need to add command that are required for the charger to properly and safely charge lithium batteries.  As is by default the high temp charge is set default to 60 C. But on Lithium batteries, charging should cut off below 0, and above 45 C thus adding simple commands to clarify and setup.